+++ src/main/scala/de/htwg/se/reversi/model/gridComponent/gridBaseImpl/gridMockImpl/Grid.scala
@@ -1,7 +1,9 @@
 package de.htwg.se.reversi.model.gridComponent.gridMockImpl
 
+import de.htwg.se.reversi.model.gridComponent.gridBaseImpl.Turn
 import de.htwg.se.reversi.model.gridComponent.{CellInterface, GridInterface}
 
+/*
 class Grid(var size:Int) extends GridInterface{
 
   size=1
@@ -13,7 +15,7 @@ class Grid(var size:Int) extends GridInterface{
   def setTurnRC(playerId: Int, row: Int, col: Int): GridInterface = this
   def reset(row:Int, col:Int): GridInterface = this
   def createNewGrid: GridInterface = this
-  //def getValidTurns(playerId:Int):List[Turn]
+  def getValidTurns(playerId:Int):List[Turn] =
   //def evaluateGame():Int
   def setHighlight(turn:Turn):GridInterface = this
   def highlight(playerId: Int): GridInterface = this
@@ -24,4 +26,4 @@ class Grid(var size:Int) extends GridInterface{
 object EmptyCell extends CellInterface {
   def value: Int = 0
   def isSet: Boolean = false
-}
+}*/
